Create balance
This contract allows the ownership owner to set the interest rate for borrowing and provides the ability to query the contract balance. Please note that the interest rate here is expressed as a percentage per year, that is, the annual interest rate. In actual applications, more refined calculations and parameter configurations may be required